+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -4,10 +4,9 @@ CHANGELOG
 2.2.0
 -----
 
- * added Symfony\Component\HttpKernel\EventListener\RouterProxyListener
  * added Symfony\Component\HttpKernel\UriSigner
  * added Symfony\Component\HttpKernel\SubRequestRenderer and rendering strategies (in Symfony\Component\HttpKernel\RenderingStrategy)
- * added Symfony\Component\HttpKernel\EventListener\RouterProxyListener
+ * added Symfony\Component\HttpKernel\EventListener\SubRequestListener
  * added Symfony\Component\HttpKernel\DependencyInjection\ContainerAwareHttpKernel
  * added ControllerReference to create reference of Controllers (used in the SubRequestRenderer class)
  * [BC BREAK] renamed TimeDataCollector::getTotalTime() to